Move old CLI code to 'legacy' file.
